+++ sdk-registration-infinity/src/test/kotlin/com/pexip/sdk/registration/infinity/internal/TestRegistrationStep.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Pexip AS
+ * Copyright 2022-2023 Pexip AS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,6 +24,8 @@ import com.pexip.sdk.api.infinity.RequestRegistrationTokenResponse
 import com.pexip.sdk.api.infinity.Token
 
 internal abstract class TestRegistrationStep : InfinityService.RegistrationStep {
+
+    override val requestBuilder: InfinityService.RequestBuilder get() = TODO()
 
     override fun requestToken(
         username: String,