+++ CHANGELOG.md
@@ -7,6 +7,12 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- `InfinityService.RequestBuilder.infinityService`,`InfinityService.ConferenceStep.requestBuilder`,
+  `InfinityService.RegistrationStep.requestBuilder`, `InfinityService.ParticipantStep.conferenceStep`,
+  `InfinityService.CallStep.participantStep` that give access to previous steps
+
 ### Changed
 
 - Kotlin to 1.9.10